Improve center frequency search algorithm
The algorith now uses the neighbour of the highest peak instead of the
second highest peak. This will avoid getting messy values when some
harmonic appears with a high energy level. There is a trade-off here
since a harmonic with an energy level higher than 50% of the fundamental
frequency will be detected instead of the original signal. This is OK
for the smart screen scenario since the center frequency is expected
within a short range, but may not be useful for other applications.